[counters] Move helper code to new file
- Most other helper code is in root level modules, this was in
  utils/__init__ for some reason.
